backend/drone-service/requirements.txt
grpcio
grpcio-tools
protobuf
fastapi
uvicorn
websockets
aiohttp
asyncpg
python-dotenv
pydantic
pydantic-settings
prometheus-client
pika
aio-pika
minio
pytest
pytest-asyncio
pytest-cov
pytest-mock